Complete cleanup by moving everything in v4 up
